light-client: Guard rpc calls with timeout
During the move to async io the block_on implementation was dropped,
which relied on tokio. In said implementation a timeout was enforced to
limit the duration of the execution of the given task. To not
permanently loose that guard rail this change introduces a timeout
implementation based on the futures-timer[0] crate. It supports a wasm
environment, which hasn't been validated or assessed yet but should give
reasonable confidence that it might work in a `no_std` environment,
q.e.d.

[0] https://crates.io/crates/futures-timer

Signed-off-by: xla <self@xla.is>